Add support for authentication via Unix socket
Just thought it'd be cool to be able to authenticate with mysql via a unix socket. A unix socket checks the user that's trying to access it and does authentication that way.

Laravel has support built in already, I just needed to define it here to get it working on my install.