Update egg-teamspeak3-server.json
Allow to users to change filetransfer port, telnet port and ssh port.
I also put the telnet port to the default one and not the voice port.
I saw one problem which I don't know to fix.
For the change of the filestransfer port you need to give an ip. So I put 0.0.0.0
Normally Teamspeak binds `2019-03-22 11:47:57.590315|INFO |VirtualServer |1 |listening on 0.0.0.0:9988, [::]:9988`
But i didn't find a way to give the second one with it. 

Log that you can see it's is starting with custom ports.
```
[Pterodactyl Daemon] Server marked as STARTING
[Pterodactyl Daemon] Checking size of server data directory...
[Pterodactyl Daemon] Disk Usage: 18M / 0M
[Pterodactyl Daemon] Ensuring file permissions.
[Pterodactyl Daemon] Running server preflight.
[Pterodactyl Daemon] Starting server container.
2019-03-22 11:54:14.502446|INFO | | |myTeamSpeak identifier revocation list was downloaded successfully - all related features are activated
2019-03-22 11:54:14.968633|INFO | | |Puzzle precompute time: 521
2019-03-22 11:54:14.968926|INFO |FileManager | |listening on 0.0.0.0:30034
[Pterodactyl Daemon] Server marked as ON
2019-03-22 11:54:14.979194|INFO |VirtualServer |1 |listening on 0.0.0.0:9988, [::]:9988
2019-03-22 11:54:14.979486|INFO |Query | |listening for query on 0.0.0.0:10012, [::]:10012
2019-03-22 11:54:14.979806|INFO |Query | |listening for query ssh on 0.0.0.0:10023, [::]:10023
2019-03-22 11:54:14.979941|INFO |CIDRManager | |updated query_ip_whitelist ips: 127.0.0.1/32, ::1/128,
```