Update egg-teamspeak3-server.json
Updated installationsscript and the Version field. version can now be empty which is the default. Users can still manually get an outdated one.
If the version field is null i extract the version from offical source via jq. It's a json and it's the easiest way to retrieve the downloadlink and the version it.
if version is set, then it's the old way with hardcoded link.