I hate frontend.
In this commit, we venture with Fallen. A young lad who hates frontend development. During his time working on fixing an issue with Jexactyl, he battles his worst enemy: frontend! Just as he thought it was working, BAM. Frontend knocks him down to his knees. So he gets a little irritated, and decides to go to the backend to solve the problem. With a few modifications and the installation of a supercharger 6900+ he manages to accomplish some of what he needs. Then he comes back stronger then ever to attack frontend with all his might. He swings his lines of code at frontend and kills it swiftly and proudly. Standing proud with victory, Fallen proceeds to make the medications he needs to the frontend and all is well is the land of Jexactyl. The end.

Fixed default nest and egg variables so they no longer default to 1 and default to the first nest and egg in the database.